requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
SQLAlchemy>=2.0
alembic
psycopg2-binary

